Fileselection refactor working 2
