+++ frontend/src/components/Import/LocalFileImport/FileSelection.tsx
@@ -2,62 +2,7 @@ import { forwardRef, useImperativeHandle, useRef, useState, FC } from 'react';
 import { Upload } from 'lucide-react';
 import type { LogSourceProvider, LogSourceProviderService } from '../types';
 import { useImportStore } from '../../../stores/useImportStore';
-
-
-export const FileSelectionService : LogSourceProviderService = {
-  name: "File",
-
-  handleFilePreview: async (file, onPreviewReadyCallback) => {
-
-    const reader = new FileReader();
-    reader.onload = async (e) => {
-      const text = e.target?.result as string;
-      const lines = text.split('\n').slice(0, 100);
-     // TBD check if the file is binary or has no valid delimiter
-      onPreviewReadyCallback(lines);
-      reader.abort();
-    };
-    reader.readAsText(file);
-  },
-  handleFileImport: (filename, filehandle, chunkSize, callback) => {  
-    // process data
-
-    let currentIndex = 0;
-    return new Promise((resolve, reject) => {
-      console.log("Importing file:", filename, "with handle:", filehandle);
-      if (filename && filehandle) {
-        const reader = new FileReader();
-        reader.onload = async (e) => {
-          const text = e.target?.result as string;
-          const lines = text.split('\n');
-          let totalLines = lines.length;
-
-          const processChunk = async () => {
-            const chunk = lines.slice(currentIndex, currentIndex + chunkSize);
-            await callback(chunk, totalLines, () => {
-              currentIndex += chunkSize;
-              if (currentIndex < totalLines) {
-                processChunk();
-              } else {
-                resolve();
-              }
-            });
-          };  
-          processChunk();
-        };
-        reader.onerror = (e) => {
-          reject(new Error("Error reading file"));
-        };
-        reader.readAsText(filehandle);
-      } else {
-        reject(new Error("No filename or filehandle provided"));
-      }
-    });
-  },
-};
-
-
-
+import { FileSelectionService } from './FileSelectionService';
 
 // Forward ref 
 const FileSelection = forwardRef<{}, LogSourceProvider>(({   
@@ -68,17 +13,6 @@ const FileSelection = forwardRef<{}, LogSourceProvider>(({
     const fileInputRef = useRef<HTMLInputElement>(null);  
     const { error, setMetadata, setSelectedFileName, setSelectedFileHandle, setFilePreviewBuffer } = useImportStore();
     const [pendingResolve, setPendingResolve] = useState<(() => void) | null>(null);
-
-    // Implement the LogSourceProviderRef interface
-    useImperativeHandle(ref, () => ({
-      handleImport: async (chunkSize, callback) => {
-        // Implementation
-      },
-      getName: () => {
-        // Return the filename of the selected file
-        return fileInputRef.current?.files?.[0]?.name || 'Unknown file';
-      }
-    }));
 
     const handleFileSelect = async (event: React.ChangeEvent<HTMLInputElement>) => {
       const file = event.target.files?.[0];
