+++ frontend/src/components/Import/UploadSteps/index.ts
@@ -1,10 +1,10 @@
 // Export components from this directory
-export { default as FileSelection } from './FileSelection';
+export { default as FileSelection } from '../LocalFileImport/FileSelection';
 export { default as ImportConfirm, SuccessSummary } from './ImportConfirmStep';
 export { default as LogSourceSelectionStep } from './LogSourceSelectionStep';
 export { default as SourceSelection } from './SourceSelection';
 
 
 // Re-export CloudWatch components from their new location
-export { CloudWatchSelection } from '@/components/CloudWatch';
-export type { CloudWatchSelectionRef } from '@/components/CloudWatch'; 
+export { CloudWatchSelection } from '@/components/Import/CloudWatchImport';
+export type { CloudWatchSelectionRef } from '@/components/Import/CloudWatchImport';
